added pattern test cases for h alphabet detectors
